+++ NEWS.md
@@ -12,3 +12,9 @@
 # NetworkExtinction 0.1.3
 
 * Added parallel processing for `RandomExtinctions`
+
+# NetworkExtinction 0.2.1
+
+* Change function name from `degree_distribution` to `DegreeDistribution`
+* The functions `Mostconnected` and `ExtinctionOrder` are now soft deprecated in favour of `SimulateExtinctions`
+* Transformed `ExtinctionOrder` output to unify with `Mostconnected` output so that `SimulateExtinctions` has a common output